+++ src/engine/engine_support.c
@@ -56,7 +56,8 @@ const char* mjDISABLESTRING[mjNDISABLE] = {
   "Actuation",
   "Refsafe",
   "Sensor",
-  "Midphase"
+  "Midphase",
+  "Eulerdamp"
 };